requirements.txt
astroid==2.2.5
certifi==2019.9.11
chardet==3.0.4
configparser==4.0.2
Django==2.2.5
djangorestframework==3.10.3
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
pylint==2.3.1
PyMySQL==0.9.3
pytz==2019.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.3
wrapt==1.11.2
